pyproject.toml
[tool.poetry]
name = "resolve-renderbot"
version = "0.1.0"
description = "Slack bot for Resolve render notifications"
authors = ["in03 <in03@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "resolve_renderbot"}]

[tool.poetry.dependencies]
python = ">=3.7.2,<4" 
slack-sdk = "^3.19.1"
rich = "^12.6.0"
pydavinci = {git = "https://github.com/in03/pydavinci"}
python-dotenv = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
